Refactor app structure and add docs
