+++ AGENTS.md
@@ -1,0 +1,6 @@
+# Contributor Guidance
+
+- Use two spaces for indentation in all TypeScript and TSX files.
+- Keep application source in the `src/` directory and group reusable UI in `src/components`.
+- Run `npm test` (or the closest available check) before committing changes.
+- Update documentation when behavior or public APIs change.
