+++ src/components/StatCard.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+export function StatCard({ label, value, hint }: { label: string; value: string; hint?: string }) {
+  return (
+    <div className="rounded-2xl border p-4 shadow-sm">
+      <div className="text-sm text-slate-500">{label}</div>
+      <div className="text-2xl font-semibold">{value}</div>
+      {hint && <div className="text-xs text-slate-500">{hint}</div>}
+    </div>
+  );
+}
